Small bug fixes
