+++ app/src/main/java/com/appdevelopement/passinggrade/MainActivity.kt
@@ -80,9 +80,9 @@ class MainActivity : AppCompatActivity() {
         // Use middleware to add entities to database
         //initializeDatabase()
 
-        //       TeacherManagerV2.addTeacher(this)
+//          TeacherManagerV2.addTeacher(this)
         //  AddStudent.addStundent(this)
-//        AddExam.addExam(this)
+//       AddExam.addExam(this)
 //        CompetenceManager.addCompetences(this)
         initializeDatabase()
 
@@ -99,21 +99,21 @@ class MainActivity : AppCompatActivity() {
         CoroutineScope(Dispatchers.IO).launch {
             // Ensure that all middleware operations are completed in the correct sequence
 
-            // Insert teacher
-            //  val teacher = TeacherManger.addTeacher(context)
-            // val teacherId = teacher.idTeacher
+//            // Insert teacher
+//              val teacher = TeacherManger.addTeacher(context)
+//            val teacherId = teacher.idTeacher
 
-            // Insert student
+//             Insert student
 //            val student = AddStudent.addStudent(context)
 //            val studentId = student.idStudent
 
-            // Insert course
-            //           val course = CourseManager.addCourse(context)
-            //           val courseId = course.idCourse
-
-            // Insert exam
-//            val exam = AddExam.addExam(context, 43, 1, 1)
-//            val examId = exam.idExam
+//            // Insert course
+//                       val course = CourseManager.addCourse(context)
+//            val courseId = course.idCourse
+//
+//            // Insert exam
+//           val exam = AddExam.addExam(context, 1, 1, 1)
+//          val examId = exam.idExam
 
             // Insert competences
 //            CompetenceManager.addCompetences(context, examId)
