+++ app/src/main/java/com/appdevelopement/passinggrade/middelware/AddStudent.kt
@@ -15,14 +15,14 @@ import kotlinx.coroutines.withContext
 object AddStudent {
     val student = Student(
         idStudent = 0,
-        studentName = "adas Doe",
+        studentName = "Bernardo Alves",
         studentNumber = 2321,
         isGraded = false
     )
 
     val student2 = Student(
         idStudent = 0,
-        studentName = "twas Doe",
+        studentName = "Mehdi Sadghi",
         studentNumber = 123131313,
         isGraded = false
     )
