+++ app/src/main/java/com/appdevelopement/passinggrade/middelware/TeacherManger.kt
@@ -11,20 +11,20 @@ import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 
 object TeacherManger {
-//    private val teacher1 = Teacher(
-//        idTeacher = 0,
-//        dtEmail = "teacher1@example.com",
-//        dtPassword = "securePassword",
-//        dtName = "Teacher 1"
-//    )
-//
-//    suspend fun addTeacher(context: Context): Teacher {
-//        val dao = AppDatabase.getDatabase(context).teacherDao()
-//        val teacherId = withContext(IO) {
-//            dao.insertTeacher(teacher1)
-//        }
-//        return teacher1.copy(idTeacher = teacherId.toInt())
-//    }
+    private val teacher1 = Teacher(
+        idTeacher = 0,
+        dtEmail = "jan@email.com",
+        dtPassword = "JanPaw",
+        dtName = "Teacher 1"
+    )
+
+    suspend fun addTeacher(context: Context): Teacher {
+        val dao = AppDatabase.getDatabase(context).teacherDao()
+        val teacherId = withContext(IO) {
+            dao.insertTeacher(teacher1)
+        }
+        return teacher1.copy(idTeacher = teacherId.toInt())
+    }
 
     suspend fun getTeacherByEmailAndPassword(context: Context, email: String, password: String): Teacher?{
         val dao = AppDatabase.getDatabase(context).teacherDao()
