+++ app/src/main/res/layout/student_page.xml
@@ -1,12 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        xmlns:tools="http://schemas.android.com/tools"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical"
-        android:padding="15dp"
-        tools:context=".pages.PracticeMainActivity">
+                                                   xmlns:app="http://schemas.android.com/apk/res-auto"
+                                                   xmlns:tools="http://schemas.android.com/tools"
+                                                   android:layout_width="match_parent"
+                                                   android:layout_height="match_parent"
+                                                   android:orientation="vertical"
+                                                   android:padding="15dp"
+                                                   tools:context=".pages.StudentPageFragment">
 
     <!-- Profile Image -->
     <ImageButton
@@ -30,7 +30,7 @@
             app:layout_constraintTop_toTopOf="@+id/guideline4"
             app:layout_constraintEnd_toStartOf="@+id/guideline3" />
 
-    <!--  Guidelines  -->
+    <!-- Guidelines -->
     <androidx.constraintlayout.widget.Guideline
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
@@ -64,7 +64,7 @@
             android:layout_height="wrap_content"
             android:id="@+id/guideline7"
             android:orientation="vertical"
-            app:layout_constraintGuide_percent="0.50364965" />
+            app:layout_constraintGuide_percent="0.5" />
 
     <!-- CourseName -->
     <TextView
@@ -79,26 +79,13 @@
             app:layout_constraintStart_toStartOf="@+id/guideline2"
             android:layout_marginTop="30dp" />
 
-    <!--  Student Number Field  -->
-<!--    <EditText-->
-<!--            android:layout_width="165dp"-->
-<!--            android:layout_height="48dp"-->
-<!--            android:ems="10"-->
-<!--            android:id="@+id/etStudentNumberField"-->
-<!--            android:hint="Search by student number"-->
-<!--            android:textSize="@dimen/_10ssp"-->
-<!--            app:layout_constraintStart_toStartOf="@+id/guideline2"-->
-<!--            app:layout_constraintTop_toBottomOf="@+id/tvCourseName"-->
-<!--            android:layout_marginTop="20dp"-->
-<!--            android:inputType="number" />-->
-
+    <!-- Student Number Field -->
     <androidx.appcompat.widget.SearchView
             android:layout_width="165dp"
             android:layout_height="48dp"
             android:ems="10"
             android:id="@+id/searchView"
             android:inputType="number"
-
             app:iconifiedByDefault="false"
             app:searchHintIcon="@null"
             android:focusable="false"
@@ -106,13 +93,8 @@
             app:layout_constraintTop_toBottomOf="@+id/tvCourseName"
             android:layout_marginTop="20dp"
             app:queryHint="Search by number..." />
-    <!-- FilterBy Spinner  -->
 
-    <!-- Export Grades Button -->
-
-    <!-- Scan Excel Sheet Button -->
-
-
+    <!-- FilterBy Spinner -->
     <Spinner
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
@@ -124,29 +106,17 @@
             app:layout_constraintTop_toBottomOf="@+id/fabMenu"
             app:layout_constraintVertical_bias="0.852" />
 
+    <!-- Import Excel Sheet Button -->
     <Button
-            android:text="Export Grades"
+            android:id="@+id/btnImportSheet"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:id="@+id/btnCreateSheet"
-            android:backgroundTint="#0C0C0C"
-            android:textColor="#FFFFFF"
-            app:layout_constraintEnd_toStartOf="@+id/guideline3"
-            app:layout_constraintBottom_toBottomOf="parent"
-            android:layout_marginBottom="44dp"
-            android:layout_marginTop="26dp"
-            app:layout_constraintTop_toBottomOf="@+id/recyclerView" />
-
-    <Button
-            android:text="Scan Excel Sheet"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:id="@+id/btnAddCriteria"
-            app:layout_constraintStart_toStartOf="@+id/guideline2"
-            app:layout_constraintBottom_toBottomOf="parent"
-            android:layout_marginBottom="44dp"
-            android:layout_marginTop="26dp"
-            app:layout_constraintTop_toBottomOf="@+id/recyclerView" />
+            android:text="Import Excel Sheet"
+            android:layout_marginTop="84dp"
+            app:layout_constraintTop_toBottomOf="@+id/recyclerView"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintHorizontal_bias="0.149" />
 
     <LinearLayout
             android:id="@+id/llStudents"
@@ -222,8 +192,8 @@
 
     <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/recyclerView"
-            android:layout_width="340dp"
-            android:layout_height="350dp"
+            android:layout_width="339dp"
+            android:layout_height="240dp"
             app:layout_constraintStart_toStartOf="@+id/guideline2"
             app:layout_constraintEnd_toStartOf="@+id/guideline3"
             app:layout_constraintTop_toBottomOf="@+id/spnrFilterBy"
