+++ settings.gradle
@@ -1,11 +1,12 @@
 pluginManagement {
-    repositories {
-        google()
-        gradlePluginPortal()
-    }
     plugins {
         id 'com.google.devtools.ksp' version '1.6.0-1.0.2'
         id 'org.jetbrains.kotlin.android' version '1.8.0'
+        id 'androidx.navigation.safeargs.kotlin' version '2.5.3'
+    }
+    repositories {
+        gradlePluginPortal()
+        google()
     }
 }
